eliminado ficheros innecesarios y refactorizacion hecha
