+++ CuantoMeme2.0/bbdd_ejem6_mio/src/main/java/es/urjc/code/daw/FileUploadController.java
@@ -90,7 +90,7 @@ public class FileUploadController {
     		}
     	}
     	
-    	Vineta viñeta = new Vineta(titulo, desc, "/imgVinetas/"+file.getOriginalFilename());
+    	Vineta viñeta = new Vineta(titulo, desc, "/imgs/"+file.getOriginalFilename());
     	Principal p = request.getUserPrincipal();
     	User user = usuarios.findByUsername(p.getName());
     	viñeta.setAutor(user);
