--- CuantoMeme2.0/bbdd_ejem6_mio/src/main/java/es/urjc/code/daw/VinetaController.java
@@ -1,5 +1,0 @@
-package es.urjc.code.daw;
-
-public class VinetaController {
-
-}
