--- CuantoMeme2.0/bbdd_ejem6_mio/src/main/resources/static/index.css
@@ -1,35 +1,0 @@
-.img-responsive {
-    margin: auto;
-}
-
-.jumbotron {
-    /*background-color: #1abc9c;*/
-    box-shadow: 0 4px 8px 0 rgba(0, 0, 0, 0.2), 0 6px 20px 0 rgba(0, 0, 0, 0.19);
-    padding-top: 5px;
-}
-
-.jumbotron > h1 {
-    text-align: center;
-}
-.btn-circle.btn-lg {
-  width: 50px;
-  height: 50px;
-  padding: 10px 16px;
-  font-size: 18px;
-  line-height: 1.33;
-  border-radius: 30px;
-}
-.jumbotron {
-    padding-right: 150px;
-    padding-left: 5px;
-}
-.jumbotron > .img-responsive{
-    padding-bottom: 25px
-}
-
-#page-wrapper {
-    background-color: #f5f5f5;
-}
-
-
-
