--- CuantoMeme2.0/bbdd_ejem6_mio/src/main/resources/static/login.html
@@ -1,11 +1,0 @@
-<html>
-<body>
-<form action="/login">
-  <p>Saludar a :</p>
-  <input type='text' name='username'/> 
-  <input type='password' name='password'/> 
-  <input type='submit' value='Login'/>
-  <input type="hidden" name="_csrf" value="{{token}}"/>
-</form>
-</body>
-</html>
