--- CuantoMeme2.0/bbdd_ejem6_mio/src/main/resources/static/registro.html
@@ -1,14 +1,0 @@
-<html>
-<body>
-<form action="registro">
- <p>Username</p>
- <input type='text' name='username'/> 
- <p>Password</p>
-  <input type='password' name='password'/> 
-  <p>Email</p>
- <input type="email" name="email"> 
-
-  <input type='submit' value='Registrarse'/>
-</form>
-</body>
-</html>
