--- CuantoMeme2.0/bbdd_ejem6_mio/src/main/resources/static/signup.html
@@ -1,15 +1,0 @@
-<html>
-<body>
-<form action="registro" method="post">
- <p>Username</p>
- <input type='text' name='username'/> 
- <p>Password</p>
-  <input type='password' name='password'/> 
-  <p>Email</p>
- <input type="email" name="email"> 
-
-  <input type='submit' value='Registrarse'/>
-    <input type="hidden" name="_csrf" value="{{token}}"/>
-</form>
-</body>
-</html>
